requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
drf-spectacular